Moving to python3 for the fuzzer

+++ scripts/fuzz/debuglib.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
 # Copyright (C) 2014  Mate Soos
@@ -103,7 +103,7 @@ class debuglib:
         # file with ::solve()
         file_len = debuglib.file_len_no_comment(fname1)
         if num_to_add > 0:
-            nextToAdd = random.randint(1, (file_len / num_to_add) + 1)
+            nextToAdd = random.randint(1, int(file_len / num_to_add) + 1)
         else:
             nextToAdd = file_len + 1
 
@@ -131,7 +131,7 @@ class debuglib:
                     fout.write("c Solver::solve( %s )\n" % assumps)
 
                 nextToAdd = at + \
-                    random.randint(1, (file_len / num_to_add) + 1)
+                    random.randint(1, int(file_len / num_to_add) + 1)
 
             # calculate max variable
             maxvar = max(maxvar, get_max_var_from_clause(line))
